Agrega funcion cobrar

+++ app/controllers/OrderController.php
@@ -1,11 +1,22 @@
   <?php
    class OrderController extends BaseController {
-      
+    
+  private $autorizado;
+   public function __construct() {
+      $this->autorizado = (Auth::check() and Auth::user()->name == 'Ariel');
+   } 
       public function index(){
+        if(!$this->autorizado) return Redirect::to('/auth/login');
           $orders =Order::where('date','=',date("Y-m-d"))->get();
-          //$orders = Order::all( array('date','status','user_id','table_id'));
           return View::make('order.index', array('orders' => $orders));
       }
+
+    public function items($id) {
+     $order = Order::find($id, array('id','user_id','table_id','status','total'));
+     $total = $order->total;
+     return View::make('order.items', array('order' => $order, 'total'=>$total));
+     }
+      
   public function finalizar($id) {
      $order = Order::find($id, array('id'));
      $order->total=Input::get('total');
@@ -18,16 +29,13 @@
      }
 
   public function show($id) {
-     $order = Order::find($id, array('id','user_id','table_id','status'));
-     $total='0';
-     foreach($order->items as $item){
-     $price=$item->price*$item->pivot->quantity;
-     $total=$total+$item->price*$item->pivot->quantity;
-     }
-     return View::make('order.show', array('order' => $order, 'total'=>$total));
+    if(!$this->autorizado) return Redirect::to('/auth/login');
+     $order = Order::find($id, array('id','user_id','table_id','status','total'));
+     return View::make('order.show', array('order' => $order));
      }
 
      public function create() { 
+      if(!$this->autorizado) return Redirect::to('/auth/login');
       $order = new Order();
       $users = User::all(array('id','name','lastname'));
       $tables= Table::where('state','=',0)->get();
@@ -35,6 +43,7 @@
      }
 
      public function store() { 
+      if(!$this->autorizado) return Redirect::to('/auth/login');
       $respuesta = array();
  
         $reglas =  array(
@@ -54,6 +63,7 @@
             $order->user_id = Input::get('user_id');
             $order->table_id = Input::get('table_id');
             $order->status = '1';
+            $order->total='0';
             $table = Table::find($order->table_id);
             $table->state ='1';
             $table->save();
@@ -70,60 +80,21 @@
           }else{
               $order = Order::find($respuesta['id']);
               $categories = Category::all(array('id','name'));
-              $total='0';
-              return View::make('order.agregar', array('order' => $order,'categories'=>$categories, $respuesta['notice'],'total'=>$total));
+              return View::make('order.agregar', array('order' => $order,'categories'=>$categories, $respuesta['notice']));
           }
      }
 
-  public function insertarItems($id) { 
+  public function editarItems($id) { 
+    if(!$this->autorizado) return Redirect::to('/auth/login');
       $order = Order::find($id);
       $categories = Category::all(array('id','name'));
-      $total='0';
-     return View::make('order.agregar', array('order' => $order,'categories'=>$categories, 'total'=>$total));
+     return View::make('order.agregar', array('order' => $order,'categories'=>$categories));
      }
 
-    public function eliminarItems($id) { 
-      $order = Order::find(Input::get('order_id'));
-      $item = Item::find($id);
-      $order->items()->detach($item);
-      $respuesta = array();
-      $respuesta['notice']="El item ha sido eliminado correctamente";
-      $categories = Category::all(array('id','name'));
-      $total='0';
-      return Redirect::to('orders/agregar/'.$order->id)->with('notice', $respuesta['notice']);
-     }
-
-  public function guardarItems() { 
-    $order = Order::find(Input::get('order_id'));
-    $reglas =  array(
-      'item_id' => 'required',
-      'quantity'=>'required'
-        );
-    $input = Input::all();
-    $validator = Validator::make($input, $reglas);
-  if ($validator->fails()){
-            $respuesta['notice'] = $validator;
-            $respuesta['error']   = true;
-            return Redirect::to('orders/agregar/'.$order->id)->withErrors($respuesta['notice'] )->withInput();
-        }
-  else{
-  $item = Item::find(Input::get('item_id'));
-  $quantity = Input::get('quantity');
-  $order->items()->attach($item, array('quantity'=>$quantity));
-  $respuesta = array();
-  $respuesta['notice']="El item se agrego correctamente";
-      $categories = Category::all(array('id','name'));
-      $total='0';
-      return Redirect::to('orders/agregar/'.$order->id)->with('notice', $respuesta['notice']);
-     }
-   }
 public function agregarItems() {
-  $order = Order::find(Input::get('order_id'));
-      $registerData = array(
-          'item_id' => Input::get('item_id'),
-          'quantity' => Input::get('quantity')
-      );
-          
+  if(!$this->autorizado) return Redirect::to('/auth/login');
+  $order = Order::find(Input::get('order_id'), array('id','total'));
+  if (Input::get('edit') == "agregar") {  
       $rules = array(
           'item_id' => 'required',
           'quantity' => 'required|min:1'
@@ -144,39 +115,70 @@ public function agregarItems() {
       )); 
           //en otro caso ingresamos al usuario en la tabla usuarios
       }else{
-          $item = Item::find(Input::get('item_id'));
+          $item = Item::find(Input::get('item_id'), array('id','price'));
           $quantity = Input::get('quantity');
-          $order->items()->attach($item, array('quantity'=>$quantity));
+          $total = $item->price * $quantity;
+          $order->total= $order->total + $total;
+          $order->save();
+          $order->items()->attach($item, array('quantity'=>$quantity, 'price'=>$total));
           return Response::json(array(
             'success'     =>  true,
             'message'     =>  'Se agrego el item correctamente'
         ));
       }
+    }
+    else{      
+      $item = Item::find(Input::get('item_id'), array('id'));
+      $precio = Input::get('item_price');
+      $order->total = $order->total - $precio;
+      $order->save();
+      $order->items()->detach($item);
+      return Response::json(array(
+            'success'     =>  true,
+            'message'     =>  'Se elimino el item correctamente'
+        ));
+    }
 }
+   /* public function eliminarItems($id) { 
+      $order = Order::find(Input::get('order_id'), array('id','total'));
+      $item = Item::find($id, array('id'));
+      $precio = Input::get('item_price');
+      $order->total = $order->total - $precio;
+      $order->save();
+      $order->items()->detach($item);
+      $respuesta = array();
+      $respuesta['notice']="El item ha sido eliminado correctamente";      
+      return Redirect::to('orders/edit/'.$order->id)->with('notice', $respuesta['notice']);
+            *return Response::json(array(
+            'message'     =>  'Se elimino el item correctamente'
+        ));*
+     }*/
+
    public function cobrar($id) {
-     $order = Order::find($id, array('id','user_id','table_id'));
-     $total='0';
-     foreach($order->items as $item){
-     $price=$item->price*$item->pivot->quantity;
-     $total=$total+$item->price*$item->pivot->quantity;
-     }
-     return View::make('order.cobrar', array('order' => $order, 'total'=>$total));
+    if(!$this->autorizado) return Redirect::to('/auth/login');
+     $order = Order::find($id, array('id','user_id','table_id','total'));
+     return View::make('order.cobrar', array('order' => $order));
      }
 
     public function save($id) { 
-   $order = Order::find($id);
-   $order->status='0';
+      if(!$this->autorizado) return Redirect::to('/auth/login');
+   $order = Order::find($id, array('id','table_id', 'status'));
+   $order->status = '0';
    $table = Table::find($order->table_id);
-   $table->state='0';
-   $table->save;
+   $table->state = '0';
+   $table->save();
    $order->save();
    return Redirect::to('orders')->with('notice', 'La orden ha sido cobrada.');
    }
 
    public function destroy($id) { 
+    if(!$this->autorizado) return Redirect::to('/auth/login');
    $order = Order::find($id);
+   foreach ( $order->items as $item) {
+   $order->items()->detach($item);
+   }
    $table = Table::find($order->table_id);
-   $table->state ='0';
+   $table->state = '0';
    $table->save();
    $order->delete();
    return Redirect::to('orders')->with('notice', 'La Orden ha sido eliminada correctamente.');
